pdfwriter: update page-referring objects
The PDF writer generates a new PdfIndirectDict for each page. However,
references to the original page object can remain in the object graph,
for example as link destinations. This results in a) nonfunctional links
and b) pulling unused trees into the document (duplicating the target pages).

This patch introduces a mapping from the original page numbers to the
new ones, and updates any references to point to the new page objects.